rename scheduler to properly reflect capability
